pyproject.toml
[project]
name = "foursight-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

start = "uv run uvicorn app.main:app --reload"

dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jwt>=1.3.1",
    "matplotlib>=3.10.1",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "tensorflow>=2.19.0",
    "uvicorn>=0.34.2",
    "yfinance>=0.2.58",
]